project/back-end/core/requirements.txt
flask
redis
psycopg2
gunicorn
Flask-Caching
psycogreen
gevent
kafka-python
